SCRUM-46 add settings to profile sidebar with SOS-notifications
